+++ module-db/Databases/AlarmsDB.hpp
@@ -1,17 +1,10 @@
 // Copyright (c) 2017-2020, Mudita Sp. z.o.o. All rights reserved.
 // For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
 
-/*
- * AlarmsDB.hpp
- *
- *  Created on: 15 lip 2019
- *      Author: kuba
- */
-
 #pragma once
 #include "Database/Database.hpp"
-#include "../Tables/AlarmsTable.hpp"
-#include "../Tables/ThreadsTable.hpp"
+#include "Tables/AlarmsTable.hpp"
+#include "Tables/AlarmsStatusTable.hpp"
 
 class AlarmsDB : public Database
 {
@@ -20,6 +13,7 @@ class AlarmsDB : public Database
     ~AlarmsDB() = default;
 
     AlarmsTable alarms;
+    AlarmsStatusTable alarmStatuses;
 
     static const char *GetDBName()
     {