+++ module-services/service-cellular/tests/unittest_mmi.cpp
@@ -29,6 +29,9 @@ TEST_CASE("MMI requests")
     };
 
     std::vector<TestCase> testCases = {
+        /// USSD
+        {R"(*100*#)", R"(AT+CUSD=1,*100*#,15)", typeid(UssdRequest)},
+
         /// ImeiRequest
         {R"(*#06#)", R"(AT+GSN)", typeid(ImeiRequest)},
 
@@ -85,18 +88,9 @@ TEST_CASE("MMI requests")
         {R"(*#33*1111*12#)", R"(AT+CLCK="AO",2,"1111",12)", typeid(CallBarringRequest)}, // query with pass and BS
         {R"(**33#)", std::string(), typeid(CallBarringRequest), false},                  // bad procedure - register
         {R"(##33#)", std::string(), typeid(CallBarringRequest), false},                  // bad procedure - erasure
-        {R"(*#33*1111*17#)",
-         std::string(),
-         typeid(CallBarringRequest),
-         false}, // unsupported BS - Voice Group Call Service
-        {R"(*#33*1111*18#)",
-         std::string(),
-         typeid(CallBarringRequest),
-         false}, // unsupported BS - Voice Broadcast Service
-        {R"(*#33*1111*99#)",
-         std::string(),
-         typeid(CallBarringRequest),
-         false}, // unsupported BS - All GPRS bearer services
+        {R"(*#33*1111*17#)", std::string(), typeid(CallBarringRequest), false}, // unsupported BS - Voice Group Call
+        {R"(*#33*1111*18#)", std::string(), typeid(CallBarringRequest), false}, // unsupported BS - Voice Broadcast
+        {R"(*#33*1111*99#)", std::string(), typeid(CallBarringRequest), false}, // unsupported BS - All GPRS bearer
         {R"(*#33*1111*45#)", std::string(), typeid(CallBarringRequest), false}, // unsupported BS - random
         /// BOIC (Bar Outgoing International Calls)
         {R"(*331#)", R"(AT+CLCK="OI",1)", typeid(CallBarringRequest)},                   // lock
@@ -251,6 +245,14 @@ TEST_CASE("MMI requests")
         {R"(*#67#)", R"(AT+CCFC=1,2)", typeid(CallForwardingRequest)},
         // alternative register and on
         {R"(*21*666555444#)", R"(AT+CCFC=0,1,"666555444")", typeid(CallForwardingRequest)},
+        // optional parameters - basic service group
+        {R"(*21*666555444*16#)", R"(AT+CCFC=0,1,"666555444",129,8)", typeid(CallForwardingRequest)},
+        {R"(*21*+48666555444*19#)", R"(AT+CCFC=0,1,"+48666555444",145,5)", typeid(CallForwardingRequest)},
+        // optional parameters - time
+        {R"(*21*666555444*16*30#)", R"(AT+CCFC=0,1,"666555444",129,8,,,30)", typeid(CallForwardingRequest)},
+        {R"(*21*+48666555444*19*20#)", R"(AT+CCFC=0,1,"+48666555444",145,5,,,20)", typeid(CallForwardingRequest)},
+        // not valid - timeout exceeds maximum
+        {R"(*21*+48666555444*19*40#)", std::string(), typeid(CallForwardingRequest), false},
 
         /// PasswordRegistrationRequest
         // total incoming and outgoing service barring (empty string)
@@ -300,13 +302,13 @@ TEST_CASE("MMI requests")
         // Change PIN by PUK more than 4
         {R"(**042*00002*11113*11113#)", R"(AT+CPWD="P2","00002","11113")", typeid(PinChangeRequest)},
         // bad procedure type *
-        {R"(*042*00002*11112*11112#)", std::string(), typeid(CallRequest)},
+        {R"(*042*00002*11112*11112#)", std::string(), typeid(UssdRequest)},
         // bad procedure type ##
         {R"(##042*00002*11112*11112#)", std::string(), typeid(CallRequest)},
         // bad procedure type #
         {R"(#042*00002*11112*11112#)", std::string(), typeid(CallRequest)},
         // bad procedure type *#
-        {R"(*#042*00002*11112*11112#)", std::string(), typeid(CallRequest)},
+        {R"(*#042*00002*11112*11112#)", std::string(), typeid(UssdRequest)},
         // no password
         {R"(**042*00002**#)", std::string(), typeid(PinChangeRequest), false},
         // no password
@@ -315,8 +317,12 @@ TEST_CASE("MMI requests")
         {R"(**042**11112*#)", std::string(), typeid(PinChangeRequest), false},
         // no password
         {R"(**042**11112*#)", std::string(), typeid(PinChangeRequest), false},
-        //  password does not match
+        // password does not match
         {R"(**042*0000*1111*2222#)", std::string(), typeid(PinChangeRequest), false},
+
+        /// call
+        {R"(666555777)", std::string(), typeid(CallRequest)},
+        {R"(+48666555777)", std::string(), typeid(CallRequest)},
     };
 
     for (auto &testCase : testCases) {
@@ -326,11 +332,14 @@ TEST_CASE("MMI requests")
         INFO("Failed on testcase: " << testCase.requestString);
         REQUIRE(typeid(*request.get()).name() == testCase.expectedType.name());
         REQUIRE(request->isValid() == testCase.expectedValid);
-        if (typeid(*request.get()).name() != typeid(CallRequest).name()) {
-            REQUIRE(requestCommand == testCase.expectedCommandString);
+        if (typeid(*request.get()).name() == typeid(CallRequest).name()) {
+            REQUIRE(requestCommand == "ATD" + testCase.requestString + ";");
+        }
+        else if (typeid(*request.get()).name() == typeid(UssdRequest).name()) {
+            REQUIRE(requestCommand == "AT+CUSD=1," + testCase.requestString + ",15");
         }
         else {
-            REQUIRE(requestCommand == "ATD" + testCase.requestString + ";");
+            REQUIRE(requestCommand == testCase.expectedCommandString);
         }
     }
 }