+++ module-services/service-db/service-db/DBServiceAPI.hpp
@@ -151,14 +151,6 @@ class DBServiceAPI
     static std::unique_ptr<utils::PhoneNumber::View> GetNumberById(sys::Service *serv,
                                                                    std::uint32_t numberId,
                                                                    std::uint32_t timeout = DefaultTimeoutInMs);
-    static bool AlarmAdd(sys::Service *serv, const AlarmsRecord &rec);
-    static bool AlarmRemove(sys::Service *serv, uint32_t id);
-    static bool AlarmUpdate(sys::Service *serv, const AlarmsRecord &rec);
-    static uint32_t AlarmGetCount(sys::Service *serv);
-    static std::unique_ptr<std::vector<AlarmsRecord>> AlarmGetLimitOffset(sys::Service *serv,
-                                                                          uint32_t offset,
-                                                                          uint32_t limit);
-    static AlarmsRecord AlarmGetNext(sys::Service *serv, time_t time);
 
     static bool NotesAdd(sys::Service *serv, const NotesRecord &rec);
     static bool NotesRemove(sys::Service *serv, uint32_t id);